+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 
 ## 2023 February
 
+February 14:
+- Title screen.
+
 February 06:
 - Refactor Fire Days's game state into core and screen.
 - Refactor Fire Days's init into core and screen.